Model-Comparision/requirements.txt
torch>=2.0.0
transformers>=4.30.0
matplotlib>=3.7.0
ipython>=8.10.0
python-dotenv>=1.0.0
huggingface_hub>=0.16.4 